feat: Use stable Rust edition 2021 (#65)
This removes our use of the `generic_const_exprs` nightly feature. This
makes the API a bit more annoying by requiring an additional `const B:
usize` parameter that was previously computed as `1024 * W / <$T>::T` on
most of the safe public API methods.

I've looked on crates.io and it appears the only published crates with
depend on fastlanes-rs are Vortex and
[LiquidCache](https://github.com/XiangpengHao/liquid-cache/blob/cdaf09fa87523d614c99d7f3e514dbe900d01ca3/src/liquid_parquet/src/liquid_array/raw/bit_pack_array.rs#L8).
Both only use the `unchecked` APIs, so this should not break either of
them.

All existing tests continue to pass.

I chose Rust Edition 2021 to be maximally compatible with downstream
users, and we don't really make use for 2024 language features anyway.

---------

Signed-off-by: Andrew Duffy <andrew@a10y.dev>